refactor(skills): modularize upload dialog and fix sub-ZIP performance
Extract types, sub-components, and createSkillSubZip into dedicated
modules. Dialog reduced from 593 to 329 lines. Fix O(N*M) performance
issue where multi-skill upload re-parsed the original ZIP for each
skill — now parsed once and cached per FileEntry. Add 5 tests for
createSkillSubZip covering extraction, nested dirs, and edge cases.
