+++ ui/web/src/pages/skills/skill-upload-dialog.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef } from "react";
 import { useTranslation } from "react-i18next";
-import { Upload, CheckCircle2, XCircle, Loader2, TriangleAlert, X, Package } from "lucide-react";
+import { Upload } from "lucide-react";
 import {
   Dialog,
   DialogContent,
@@ -11,51 +11,18 @@ import {
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { validateMultiSkillZip } from "./lib/validate-skill-zip";
+import { createSkillSubZip } from "./lib/create-skill-sub-zip";
 import { uniqueId } from "@/lib/utils";
 import type { SkillUploadResponse } from "./hooks/use-skills";
+import type { FileEntry, SkillStatus } from "./lib/skill-upload-types";
+import { FileEntryBlock } from "./skill-upload-entry";
 import JSZip from "jszip";
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-/** Per-skill lifecycle status */
-type SkillStatus =
-  | "validating"
-  | "valid"        // new skill, ready to upload
-  | "unchanged"    // server returned identical hash — skipped
-  | "invalid"
-  | "uploading"
-  | "success"
-  | "warning"      // uploaded but deps_warning present
-  | "error";
-
-interface SkillEntry {
-  id: string;
-  dir: string;
-  status: SkillStatus;
-  name?: string;
-  slug?: string;
-  contentHash?: string;
-  error?: string;
-}
-
-interface FileEntry {
-  id: string;
-  file: File;
-  /** One entry per detected skill in this ZIP */
-  skills: SkillEntry[];
-}
 
 interface SkillUploadDialogProps {
   open: boolean;
   onOpenChange: (open: boolean) => void;
   onUpload: (file: File) => Promise<SkillUploadResponse>;
 }
-
-// ---------------------------------------------------------------------------
-// Component
-// ---------------------------------------------------------------------------
 
 export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadDialogProps) {
   const { t } = useTranslation("skills");
@@ -91,7 +58,6 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
           const validation = await validateMultiSkillZip(entry.file);
           const placeholderId = entry.skills[0]?.id ?? uniqueId();
           if (validation.error) {
-            // Top-level ZIP error (corrupt, too large, not a zip)
             return {
               id: entry.id,
               skills: [{ id: placeholderId, dir: "", status: "invalid" as SkillStatus, error: validation.error }],
@@ -103,16 +69,18 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
               skills: [{ id: placeholderId, dir: "", status: "invalid" as SkillStatus, error: "upload.noSkillMd" }],
             };
           }
-          const skills: SkillEntry[] = validation.skills.map((s) => ({
-            id: uniqueId(),
-            dir: s.dir,
-            status: s.valid ? ("valid" as SkillStatus) : ("invalid" as SkillStatus),
-            name: s.name,
-            slug: s.slug,
-            contentHash: s.contentHash,
-            error: s.error,
-          }));
-          return { id: entry.id, skills };
+          return {
+            id: entry.id,
+            skills: validation.skills.map((s) => ({
+              id: uniqueId(),
+              dir: s.dir,
+              status: s.valid ? ("valid" as SkillStatus) : ("invalid" as SkillStatus),
+              name: s.name,
+              slug: s.slug,
+              contentHash: s.contentHash,
+              error: s.error,
+            })),
+          };
         } catch {
           return {
             id: entry.id,
@@ -135,11 +103,10 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
   };
 
   // ---------------------------------------------------------------------------
-  // Upload
+  // Upload — parse each ZIP once, reuse across all skills in that file
   // ---------------------------------------------------------------------------
 
   const handleSubmit = async () => {
-    // Gather all actionable skills across all file entries
     const actionable = entries.flatMap((e) =>
       e.skills
         .filter((s) => s.status === "valid")
@@ -149,8 +116,11 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
 
     setUploading(true);
 
+    // Cache parsed JSZip instances keyed by FileEntry id — avoids re-parsing
+    // the same ZIP blob for every skill in a multi-skill archive (O(N) vs O(N*M)).
+    const parsedZips = new Map<string, JSZip>();
+
     for (const { fileEntry, skill } of actionable) {
-      // Mark this skill as uploading
       setEntries((prev) =>
         prev.map((e) =>
           e.id === fileEntry.id
@@ -160,26 +130,24 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
       );
 
       try {
-        // For multi-skill ZIPs extract just this skill's files into a sub-ZIP.
-        // Single-skill ZIPs (dir === "") upload the original file directly.
-        const uploadFile =
-          skill.dir && fileEntry.skills.length > 1
-            ? await createSkillSubZip(fileEntry.file, skill.dir)
-            : fileEntry.file;
+        let uploadFile: File;
+        if (skill.dir && fileEntry.skills.length > 1) {
+          // Parse the ZIP once per FileEntry; reuse on subsequent skills
+          if (!parsedZips.has(fileEntry.id)) {
+            parsedZips.set(fileEntry.id, await JSZip.loadAsync(fileEntry.file));
+          }
+          uploadFile = await createSkillSubZip(parsedZips.get(fileEntry.id)!, skill.dir);
+        } else {
+          uploadFile = fileEntry.file;
+        }
 
         const result = await onUpload(uploadFile);
 
-        // Handle "unchanged" response (content hash matched on server)
         if (result.status === "unchanged") {
           setEntries((prev) =>
             prev.map((e) =>
               e.id === fileEntry.id
-                ? {
-                    ...e,
-                    skills: e.skills.map((s) =>
-                      s.id === skill.id ? { ...s, status: "unchanged" as SkillStatus } : s,
-                    ),
-                  }
+                ? { ...e, skills: e.skills.map((s) => s.id === skill.id ? { ...s, status: "unchanged" as SkillStatus } : s) }
                 : e,
             ),
           );
@@ -260,7 +228,7 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
   };
 
   // ---------------------------------------------------------------------------
-  // Derived counts (operate on skill level, not file level)
+  // Derived counts (skill level, not file level)
   // ---------------------------------------------------------------------------
 
   const allSkills = entries.flatMap((e) => e.skills);
@@ -358,236 +326,4 @@ export function SkillUploadDialog({ open, onOpenChange, onUpload }: SkillUploadD
       </DialogContent>
     </Dialog>
   );
-}
-
-// ---------------------------------------------------------------------------
-// FileEntryBlock — renders a ZIP file with its skill rows
-// ---------------------------------------------------------------------------
-
-function FileEntryBlock({
-  entry,
-  onRemove,
-  uploading,
-  t,
-}: {
-  entry: FileEntry;
-  onRemove: () => void;
-  uploading: boolean;
-  t: (key: string, opts?: Record<string, unknown>) => string;
-}) {
-  const isMulti = entry.skills.length > 1;
-  const sizeKB = (entry.file.size / 1024).toFixed(1);
-  const isValidating = entry.skills.some((s) => s.status === "validating");
-
-  if (!isMulti) {
-    // Single-skill: render as before — one row, ZIP filename as subtitle
-    const skill = entry.skills[0]!;
-    return (
-      <SkillRow
-        skill={skill}
-        subtitle={skill.name ? entry.file.name : undefined}
-        primaryLabel={skill.name || entry.file.name}
-        sizeKB={sizeKB}
-        showSize
-        onRemove={onRemove}
-        uploading={uploading}
-        t={t}
-      />
-    );
-  }
-
-  // Multi-skill: group header + individual skill rows
-  return (
-    <div className="rounded-md border overflow-hidden">
-      {/* ZIP group header */}
-      <div className="flex items-center gap-2 bg-muted/40 px-3 py-1.5 text-xs text-muted-foreground">
-        <Package className="h-3.5 w-3.5 shrink-0" />
-        <span className="flex-1 truncate font-medium">{entry.file.name}</span>
-        <span className="shrink-0">{sizeKB} KB</span>
-        {isValidating ? null : (
-          <span className="shrink-0">
-            {t("upload.multiDetected", { count: entry.skills.length })}
-          </span>
-        )}
-        {!uploading && (
-          <button
-            type="button"
-            aria-label={t("upload.remove")}
-            onClick={(e) => { e.stopPropagation(); onRemove(); }}
-            className="shrink-0 rounded-sm p-0.5 text-muted-foreground hover:text-foreground"
-          >
-            <X className="h-3 w-3" />
-          </button>
-        )}
-      </div>
-
-      {/* Individual skill rows — indented */}
-      <div className="flex flex-col divide-y">
-        {entry.skills.map((skill) => (
-          <SkillRow
-            key={skill.id}
-            skill={skill}
-            primaryLabel={skill.name || skill.dir || skill.slug || "…"}
-            subtitle={skill.dir || undefined}
-            showSize={false}
-            onRemove={undefined}
-            uploading={uploading}
-            t={t}
-            indent
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// SkillRow — single skill entry row
-// ---------------------------------------------------------------------------
-
-function SkillRow({
-  skill,
-  primaryLabel,
-  subtitle,
-  sizeKB,
-  showSize,
-  onRemove,
-  uploading,
-  t,
-  indent = false,
-}: {
-  skill: SkillEntry;
-  primaryLabel: string;
-  subtitle?: string;
-  sizeKB?: string;
-  showSize: boolean;
-  onRemove?: () => void;
-  uploading: boolean;
-  t: (key: string, opts?: Record<string, unknown>) => string;
-  indent?: boolean;
-}) {
-  const canRemove = !uploading && skill.status !== "uploading" && skill.status !== "success" && onRemove;
-
-  return (
-    <div className={`flex items-center gap-2 px-3 py-2 text-sm ${indent ? "pl-6" : "rounded-md border"}`}>
-      <SkillStatusIcon status={skill.status} />
-      <div className="flex-1 min-w-0">
-        <div className="flex items-center gap-1.5">
-          <span className="truncate font-medium">{primaryLabel}</span>
-          <SkillBadge status={skill.status} t={t} />
-          {showSize && sizeKB && (
-            <span className="shrink-0 text-xs text-muted-foreground">{sizeKB} KB</span>
-          )}
-        </div>
-        <SkillSubtitle skill={skill} subtitle={subtitle} t={t} />
-      </div>
-      {canRemove && (
-        <button
-          type="button"
-          aria-label={t("upload.remove")}
-          onClick={(e) => { e.stopPropagation(); onRemove(); }}
-          className="shrink-0 rounded-sm p-1 text-muted-foreground hover:text-foreground"
-        >
-          <X className="h-3.5 w-3.5" />
-        </button>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Sub-components
-// ---------------------------------------------------------------------------
-
-function SkillSubtitle({
-  skill,
-  subtitle,
-  t,
-}: {
-  skill: SkillEntry;
-  subtitle?: string;
-  t: (key: string, opts?: Record<string, unknown>) => string;
-}) {
-  if (skill.status === "invalid" || skill.status === "error") {
-    return (
-      <p className="text-xs text-destructive truncate">
-        {skill.error ? t(skill.error) : t("upload.failed")}
-      </p>
-    );
-  }
-  if (skill.status === "warning") {
-    return <p className="text-xs text-amber-600 truncate">{skill.error ?? t("upload.failed")}</p>;
-  }
-  if (skill.status === "validating") {
-    return <p className="text-xs text-muted-foreground">{t("upload.validating")}</p>;
-  }
-  if (skill.status === "unchanged") {
-    return (
-      <p className="text-xs text-muted-foreground truncate">
-        {t("upload.skillUnchanged")}
-      </p>
-    );
-  }
-  if (subtitle && skill.status !== "success") {
-    return <p className="text-xs text-muted-foreground truncate">{subtitle}</p>;
-  }
-  return null;
-}
-
-/** Colored badge for NEW / UNCHANGED / ERROR states */
-function SkillBadge({ status, t }: { status: SkillStatus; t: (key: string) => string }) {
-  const base = "shrink-0 text-[10px] font-semibold uppercase px-1.5 py-0.5 rounded";
-  switch (status) {
-    case "valid":
-      return <span className={`${base} bg-green-50 text-green-700 border border-green-200`}>{t("upload.new")}</span>;
-    case "unchanged":
-      return <span className={`${base} bg-purple-50 text-purple-700 border border-purple-200`}>{t("upload.unchanged")}</span>;
-    case "invalid":
-    case "error":
-      return <span className={`${base} bg-red-50 text-red-700 border border-red-200`}>{t("upload.failed")}</span>;
-    default:
-      return null;
-  }
-}
-
-function SkillStatusIcon({ status }: { status: SkillStatus }) {
-  switch (status) {
-    case "validating":
-    case "uploading":
-      return <Loader2 className="h-4 w-4 shrink-0 animate-spin text-muted-foreground" />;
-    case "valid":
-      return <CheckCircle2 className="h-4 w-4 shrink-0 text-primary" />;
-    case "unchanged":
-      return <CheckCircle2 className="h-4 w-4 shrink-0 text-muted-foreground" />;
-    case "success":
-      return <CheckCircle2 className="h-4 w-4 shrink-0 text-green-600" />;
-    case "warning":
-      return <TriangleAlert className="h-4 w-4 shrink-0 text-amber-600" />;
-    case "invalid":
-    case "error":
-      return <XCircle className="h-4 w-4 shrink-0 text-destructive" />;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// createSkillSubZip — extract one skill's directory into a standalone ZIP
-// ---------------------------------------------------------------------------
-
-/** Extract files under `dir/` from the original ZIP, strip the prefix, return a new File */
-async function createSkillSubZip(originalFile: File, dir: string): Promise<File> {
-  const zip = await JSZip.loadAsync(originalFile);
-  const sub = new JSZip();
-  const prefix = dir + "/";
-
-  for (const [path, entry] of Object.entries(zip.files)) {
-    if (path.startsWith(prefix) && !entry.dir) {
-      const relativePath = path.slice(prefix.length);
-      if (relativePath) {
-        sub.file(relativePath, await entry.async("blob"));
-      }
-    }
-  }
-
-  const blob = await sub.generateAsync({ type: "blob" });
-  return new File([blob], `${dir}.zip`, { type: "application/zip" });
 }
